put step: input dir: validate git repo
PCI-2587
