+++ resource/resource_test.go
@@ -262,129 +262,6 @@ Cogito SOURCE configuration:
 	}
 }
 
-func TestCheckRepoDirSuccess(t *testing.T) {
-	const wantOwner = "smiling"
-	const wantRepo = "butterfly"
-
-	testCases := []struct {
-		name    string
-		dir     string // repoURL to put in file <dir>/.git/config
-		repoURL string
-	}{
-		{
-			name:    "repo with good SSH remote",
-			dir:     "a-repo",
-			repoURL: sshRemote(wantOwner, wantRepo),
-		},
-		{
-			name:    "repo with good HTTPS remote",
-			dir:     "a-repo",
-			repoURL: httpsRemote(wantOwner, wantRepo),
-		},
-		{
-			name:    "repo with good HTTP remote",
-			dir:     "a-repo",
-			repoURL: httpRemote(wantOwner, wantRepo),
-		},
-		{
-			name:    "PR resource but with basic auth in URL (see PR #46)",
-			dir:     "a-repo",
-			repoURL: "https://x-oauth-basic:ghp_XXX@github.com/smiling/butterfly.git",
-		},
-	}
-
-	for _, tc := range testCases {
-		inDir := setup(t, tc.dir, tc.repoURL, "dummySHA", "dummyHead")
-
-		t.Run(tc.name, func(t *testing.T) {
-			err := checkRepoDir(filepath.Join(inDir, tc.dir), wantOwner, wantRepo)
-
-			if err != nil {
-				t.Fatalf("\nhave: %s\nwant: <no error>", err)
-			}
-		})
-	}
-}
-
-func TestCheckRepoDirFailure(t *testing.T) {
-	const wantOwner = "smiling"
-	const wantRepo = "butterfly"
-
-	testCases := []struct {
-		name        string
-		dir         string
-		repoURL     string // repoURL to put in file <dir>/.git/config
-		wantErrWild string // wildcard matching
-	}{
-		{
-			name:        "dir is not a repo",
-			dir:         "not-a-repo",
-			repoURL:     "dummyurl",
-			wantErrWild: `parsing .git/config: open */not-a-repo/.git/config: no such file or directory`,
-		},
-		{
-			name:        "bad file .git/config",
-			dir:         "repo-bad-git-config",
-			repoURL:     "dummyurl",
-			wantErrWild: `.git/config: key [remote "origin"]/url: not found`,
-		},
-		{
-			name:    "repo with unrelated HTTPS remote",
-			dir:     "a-repo",
-			repoURL: httpsRemote("owner-a", "repo-a"),
-			wantErrWild: `the received git repository is incompatible with the Cogito configuration.
-
-Git repository configuration (received as 'inputs:' in this PUT step):
-      url: https://github.com/owner-a/repo-a.git
-    owner: owner-a
-     repo: repo-a
-
-Cogito SOURCE configuration:
-    owner: smiling
-     repo: butterfly`,
-		},
-		{
-			name:    "repo with unrelated SSH remote or wrong source config",
-			dir:     "a-repo",
-			repoURL: sshRemote("owner-a", "repo-a"),
-			wantErrWild: `the received git repository is incompatible with the Cogito configuration.
-
-Git repository configuration (received as 'inputs:' in this PUT step):
-      url: git@github.com:owner-a/repo-a.git
-    owner: owner-a
-     repo: repo-a
-
-Cogito SOURCE configuration:
-    owner: smiling
-     repo: butterfly`,
-		},
-		{
-			name:        "invalid git pseudo URL in .git/config",
-			dir:         "a-repo",
-			repoURL:     "foo://bar",
-			wantErrWild: `.git/config: remote: invalid git URL foo://bar: invalid scheme: foo`,
-		},
-	}
-
-	for _, tc := range testCases {
-		inDir := setup(t, tc.dir, tc.repoURL, "dummySHA", "dummyHead")
-
-		t.Run(tc.name, func(t *testing.T) {
-			err := checkRepoDir(filepath.Join(inDir, tc.dir), wantOwner, wantRepo)
-
-			if err == nil {
-				t.Fatalf("\nhave: <no error>\nwant: %s", tc.wantErrWild)
-			}
-
-			have := err.Error()
-			if !wild.Match(tc.wantErrWild, have, false) {
-				diff := cmp.Diff(tc.wantErrWild, have)
-				t.Fatalf("error msg wildcard mismatch: (-want +have):\n%s", diff)
-			}
-		})
-	}
-}
-
 func TestGitGetCommitSuccess(t *testing.T) {
 	const wantSHA = "af6cd86e98eb1485f04d38b78d9532e916bbff02"
 	const defHead = "ref: refs/heads/a-branch-FIXME"
@@ -465,167 +342,6 @@ func TestGitGetCommitFailure(t *testing.T) {
 			if !wild.Match(tc.wantErrWild, have, false) {
 				diff := cmp.Diff(tc.wantErrWild, have)
 				t.Fatalf("error msg wildcard mismatch: (-want +have):\n%s", diff)
-			}
-		})
-	}
-}
-
-// setup creates a temporary directory by rendering the templated contents of dir
-// (assumed to be below testdata) with values from the remaining arguments and returns
-// the path to the directory.
-// The temporary directory is registered for removal via t.Cleanup.
-// If any operation fails, setup terminates the test by calling t.Fatal.
-func setup(
-	t *testing.T,
-	testDir string,
-	repoURL string,
-	commitSHA string,
-	head string,
-) string {
-	inDir, err := os.MkdirTemp("", "cogito-test-")
-	if err != nil {
-		t.Fatal("setup: MkdirTemp", err)
-	}
-
-	t.Cleanup(func() {
-		if err := os.RemoveAll(inDir); err != nil {
-			t.Fatal("setup: cleanup: RemoveAll:", err)
-		}
-	})
-
-	// Prepare the template data.
-	tdata := make(help.TemplateData)
-	tdata["repo_url"] = repoURL
-	tdata["commit_sha"] = commitSHA
-	tdata["head"] = head
-	tdata["branch_name"] = "a-branch-FIXME"
-
-	// Copy the testdata over
-	err = help.CopyDir(inDir, filepath.Join("testdata", testDir), help.DotRenamer, tdata)
-	if err != nil {
-		t.Fatal("CopyDir:", err)
-	}
-
-	return inDir
-}
-
-// sshRemote returns a github SSH URL
-func sshRemote(owner, repo string) string {
-	return fmt.Sprintf("git@github.com:%s/%s.git", owner, repo)
-}
-
-// httpsRemote returns a github HTTPS URL
-func httpsRemote(owner, repo string) string {
-	return fmt.Sprintf("https://github.com/%s/%s.git", owner, repo)
-}
-
-// httpRemote returns a github HTTP URL
-func httpRemote(owner, repo string) string {
-	return fmt.Sprintf("http://github.com/%s/%s.git", owner, repo)
-}
-
-func TestParseGitPseudoURLSuccess(t *testing.T) {
-	testCases := []struct {
-		name   string
-		inURL  string
-		wantGU gitURL
-	}{
-		{
-			name:  "valid SSH URL",
-			inURL: "git@github.com:Pix4D/cogito.git",
-			wantGU: gitURL{
-				URL: &url.URL{
-					Scheme: "ssh",
-					User:   url.User("git"),
-					Host:   "github.com",
-					Path:   "/Pix4D/cogito.git",
-				},
-				Owner: "Pix4D",
-				Repo:  "cogito",
-			},
-		},
-		{
-			name:  "valid HTTPS URL",
-			inURL: "https://github.com/Pix4D/cogito.git",
-			wantGU: gitURL{
-				URL: &url.URL{
-					Scheme: "https",
-					Host:   "github.com",
-					Path:   "/Pix4D/cogito.git",
-				},
-				Owner: "Pix4D",
-				Repo:  "cogito",
-			},
-		},
-		{
-			name:  "valid HTTP URL",
-			inURL: "http://github.com/Pix4D/cogito.git",
-			wantGU: gitURL{
-				URL: &url.URL{
-					Scheme: "http",
-					Host:   "github.com",
-					Path:   "/Pix4D/cogito.git",
-				},
-				Owner: "Pix4D",
-				Repo:  "cogito",
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			gitUrl, err := parseGitPseudoURL(tc.inURL)
-
-			if err != nil {
-				t.Fatalf("\nhave: %s\nwant: <no error>", err)
-			}
-			if diff := cmp.Diff(tc.wantGU, gitUrl, cmp.Comparer(
-				func(x, y *url.Userinfo) bool {
-					return x.String() == y.String()
-				})); diff != "" {
-				t.Errorf("gitURL: (-want +have):\n%s", diff)
-			}
-		})
-	}
-}
-
-func TestParseGitPseudoURLFailure(t *testing.T) {
-	testCases := []struct {
-		name    string
-		inURL   string
-		wantErr string
-	}{
-		{
-			name:    "totally invalid URL",
-			inURL:   "hello",
-			wantErr: "invalid git URL hello: missing scheme",
-		},
-		{
-			name:    "invalid SSH URL",
-			inURL:   "git@github.com/Pix4D/cogito.git",
-			wantErr: "invalid git SSH URL git@github.com/Pix4D/cogito.git: want exactly one ':'",
-		},
-		{
-			name:    "invalid HTTPS URL",
-			inURL:   "https://github.com:Pix4D/cogito.git",
-			wantErr: `parse "https://github.com:Pix4D/cogito.git": invalid port ":Pix4D" after host`,
-		},
-		{
-			name:    "invalid HTTP URL",
-			inURL:   "http://github.com:Pix4D/cogito.git",
-			wantErr: `parse "http://github.com:Pix4D/cogito.git": invalid port ":Pix4D" after host`,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			_, err := parseGitPseudoURL(tc.inURL)
-
-			if err == nil {
-				t.Fatalf("have: <no error>; want: %v", tc.wantErr)
-			}
-			if diff := cmp.Diff(tc.wantErr, err.Error()); diff != "" {
-				t.Errorf("error message mismatch: (-want +have):\n%s", diff)
 			}
 		})
 	}
